TST: add som RootTreeView setup tests, move an existing one around
