+++ superscore/tests/test_widgets.py
@@ -5,9 +5,8 @@ from uuid import uuid4
 import pytest
 from pytestqt.qtbot import QtBot
 
-from superscore.model import Collection, Root
+from superscore.model import Collection
 from superscore.widgets.core import DataWidget
-from superscore.widgets.views import RootTree
 
 
 @pytest.mark.parametrize(
@@ -39,11 +38,3 @@ def test_collection_datawidget_bridge(
 
     qtbot.addWidget(widget1)
     qtbot.addWidget(widget2)
-
-
-def test_roottree_setup(sample_database: Root):
-    tree_model = RootTree(base_entry=sample_database)
-    root_index = tree_model.index_from_item(tree_model.root_item)
-    # Check that the entire tree was created
-    assert tree_model.rowCount(root_index) == 4
-    assert tree_model.root_item.child(3).childCount() == 3
